python/requirements.txt
faker==33.3.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
six==1.17.0
tqdm==4.67.1
typing-extensions==4.12.2
xlsxwriter==3.2.0
